requirements.txt
arxiv
pymupdf
pdfplumber
pillow==10.4.0
torch==2.6.0
torchvision==0.21.0
colpali-engine==0.3.1
transformers==4.46.3
peft==0.11.1
huggingface-hub==0.26.5
qdrant-client==1.12.1
numpy==1.26.4
streamlit
accelerate